Clean up conftest
